Fix a bug in AOT where missing binding methods for ArrayTypes[] produced a crash in the processor
In the process, extract a KeyVariableNamer from BindingVariableNamer

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=221166332
