+++ java/dagger/internal/codegen/BindingVariableNamer.java
@@ -16,88 +16,22 @@
 
 package dagger.internal.codegen;
 
-import static com.google.common.base.CaseFormat.LOWER_CAMEL;
-import static com.google.common.base.CaseFormat.UPPER_CAMEL;
-import static dagger.internal.codegen.ConfigurationAnnotations.isSubcomponentBuilder;
-import static dagger.internal.codegen.SourceFiles.protectAgainstKeywords;
-
 import dagger.model.DependencyRequest;
 import dagger.model.Key;
-import java.util.Iterator;
-import javax.lang.model.element.Element;
-import javax.lang.model.type.ArrayType;
-import javax.lang.model.type.DeclaredType;
-import javax.lang.model.type.PrimitiveType;
-import javax.lang.model.type.TypeMirror;
-import javax.lang.model.type.TypeVisitor;
-import javax.lang.model.util.SimpleTypeVisitor6;
 
 /**
  * Suggests a variable name for a type based on a {@link Binding}. Prefer
  * {@link DependencyVariableNamer} for cases where a specific {@link DependencyRequest} is present.
  */
 final class BindingVariableNamer {
-  private static final TypeVisitor<Void, StringBuilder> TYPE_NAMER =
-      new SimpleTypeVisitor6<Void, StringBuilder>() {
-        @Override
-        public Void visitDeclared(DeclaredType declaredType, StringBuilder builder) {
-          Element element = declaredType.asElement();
-          if (isSubcomponentBuilder(element)) {
-            // Most Subcomponent builders are named "Builder", so add their associated
-            // Subcomponent type so that they're not all "builderProvider{N}"
-            builder.append(element.getEnclosingElement().getSimpleName());
-          }
-          builder.append(element.getSimpleName());
-          Iterator<? extends TypeMirror> argumentIterator =
-              declaredType.getTypeArguments().iterator();
-          if (argumentIterator.hasNext()) {
-            builder.append("Of");
-            TypeMirror first = argumentIterator.next();
-            first.accept(this, builder);
-            while (argumentIterator.hasNext()) {
-              builder.append("And");
-              argumentIterator.next().accept(this, builder);
-            }
-          }
-          return null;
-        }
-
-        @Override
-        public Void visitPrimitive(PrimitiveType type, StringBuilder builder) {
-          builder.append(LOWER_CAMEL.to(UPPER_CAMEL, type.toString()));
-          return null;
-        }
-
-        @Override
-        public Void visitArray(ArrayType type, StringBuilder builder) {
-          type.getComponentType().accept(this, builder);
-          builder.append("Array");
-          return null;
-        }
-      };
-
   private BindingVariableNamer() {}
 
   static String name(Binding binding) {
-    StringBuilder builder = new StringBuilder();
-
     Key key = binding.key();
-    if (key.qualifier().isPresent()) {
-      // TODO(gak): Use a better name for fields with qualifiers with members.
-      builder.append(key.qualifier().get().getAnnotationType().asElement().getSimpleName());
-    }
-
-    TypeMirror type = typeToName(binding);
-    type.accept(TYPE_NAMER, builder);
-
-    return protectAgainstKeywords(UPPER_CAMEL.to(LOWER_CAMEL, builder.toString()));
-  }
-
-  private static TypeMirror typeToName(Binding binding) {
     if (binding instanceof ContributionBinding
         && ((ContributionBinding) binding).contributionType().equals(ContributionType.SET)) {
-      return SetType.from(binding.key().type()).elementType();
+      key = key.toBuilder().type(SetType.from(key.type()).elementType()).build();
     }
-    return binding.key().type();
+    return KeyVariableNamer.name(key);
   }
 }
