+++ javatests/dagger/internal/codegen/AheadOfTimeSubcomponentsTest.java
@@ -4286,7 +4286,7 @@ public final class AheadOfTimeSubcomponentsTest {
             "  private void initialize(final Builder builder) {",
             "    this.responseProducer =",
             "        ResponseProducerModule_ResponseFactory.create(",
-            "            getExecutorProvider(),",
+            "            getProductionImplementationExecutorProvider(),",
             "            getProductionComponentMonitorProvider(),",
             "            getResponseDependencyProducer());",
             "    this.setOfResponseProducer =",
@@ -4303,7 +4303,7 @@ public final class AheadOfTimeSubcomponentsTest {
             "    return responsesEntryPoint.get();",
             "  }",
             "",
-            "  public abstract Provider<Executor> getExecutorProvider();",
+            "  public abstract Provider<Executor> getProductionImplementationExecutorProvider();",
             "",
             "  public abstract Provider<ProductionComponentMonitor>",
             "      getProductionComponentMonitorProvider();",
@@ -4530,7 +4530,7 @@ public final class AheadOfTimeSubcomponentsTest {
             "    }",
             "",
             "    @Override",
-            "    public Provider<Executor> getExecutorProvider() {",
+            "    public Provider<Executor> getProductionImplementationExecutorProvider() {",
             "      return DaggerRoot.this.productionImplementationExecutorProvider;",
             "    }",
             "",
